+++ feature-auth/api/src/main/kotlin/domain/model/SignUpModel.kt
@@ -22,6 +22,6 @@ class SignUpModel(
     val address: AddressSignUpModel,
     val carModel: String,
     val carCategory: String,
-    val carLoadCapacity: Int,
+    val carLoadCapacity: Double,
     val carPalletCapacity: Int
 )
